app/paint/gimpconvolve.c app/paint-funcs/paint-funcs.[ch] applied a patch
2004-05-31  Sven Neumann  <sven@gimp.org>

	* app/paint/gimpconvolve.c
	* app/paint-funcs/paint-funcs.[ch]
	* app/tools/gimpiscissorstool.c: applied a patch from Philip
	Lafleur that fixes RGBA resampling in Convolve tool (bug #72878).

+++ app/paint-funcs/paint-funcs.h
@@ -349,9 +349,9 @@ void  extract_from_region                 (PixelRegion *src,
 
 void  convolve_region                     (PixelRegion         *srcR,
                                            PixelRegion         *destR,
-                                           gint                *matrix,
+                                           gfloat              *matrix,
                                            gint                 size,
-                                           gint                 divisor,
+                                           gdouble              divisor,
                                            GimpConvolutionType  mode);
 
 void  multiply_alpha_region               (PixelRegion *srcR);
